requirements.txt
dotenv
fastapi
aiofiles
pyyaml
google-generativeai
uvicorn
